review remarks
